Experimental generic icon support
