+++ errors.h
@@ -1,0 +1,15 @@
+#ifndef ERRORS_H
+#define ERRORS_H
+
+#define ERROR_GENERATING_ICON 1
+#define ERROR_IMAGE_TYPE_NOT_SUPPORTED 2
+#define ERROR_INVALID_PARAMETER 3
+#define ERROR_MISSING_REQUIRED_PARAMETER 4
+#define ERROR_GENERATING_IMAGE 5
+#define ERROR_CREATING_IMAGE_DESTINATION 6
+#define ERROR_FINALIZING_IMAGE 7
+#define ERROR_WRITING_OUTPUT 8
+#define ERROR_NOTHING_TO_GENERATE 9
+#define ERROR_INTERNAL_ERROR 127
+
+#endif /* ERRORS.H */
